Progression on UI.
git-svn-id: file:///home/michael/svn/yuni/trunk@1248 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/gfx/surface/cairo.cpp
@@ -17,7 +17,7 @@ namespace Surface
 	void Cairo::release()
 	{}
 
-	void Cairo::resize(unsigned int, unsigned int)
+	void Cairo::resize(float, float)
 	{
 	}
